stop reading events when a gap of sequenceNr is detected
This can be caused by the journal writer having failed over to a
different node, and a new batch of events becoming visible
to the query client before the old batch is. It's safe to ignore
the new events, assuming the in-between batch will eventually
become visible. We're polling, after all.